Split main.dart into logical chunks
